chore: move utils to lib
