+++ spec/resolve/natural-bem-ordering/decl-ordering.spec.js
@@ -3,7 +3,7 @@
 const test = require('ava');
 const expect = require('chai').expect;
 const resolve = require('../../../lib').resolve;
-const findIndex = require('../../utils').findIndex;
+const findIndex = require('../../../lib/utils').findIndex;
 
 test('should place block before its element', () => {
     var decl = [
