+++ spec/resolve/ordering-priority/bem-vs-recommended-deps.spec.js
@@ -3,7 +3,7 @@
 const test = require('ava');
 const expect = require('chai').expect;
 const resolve = require('../../../lib').resolve;
-const findIndex = require('../../utils').findIndex;
+const findIndex = require('../../../lib/utils').findIndex;
 
 test('should prioritise block-element natural ordering over recommended deps ordering', () => {
     var decl = [
