+++ spec/resolve/unordered/recommended-ordering.spec.js
@@ -3,7 +3,7 @@
 const test = require('ava');
 const expect = require('chai').expect;
 const resolve = require('../../../lib').resolve;
-const findIndex = require('../../utils').findIndex;
+const findIndex = require('../../../lib/utils').findIndex;
 
 test('should keep the ordering described in decl', () => {
     var decl = [
